rust: add rust libcommon-core package.
It does the binding of all <lib-common/core.h> and
add some minimal support for lstr_t and pstream_t.

The waf target is called libcommon-core-rs.

We need to add a compile group for `libcommon-minimal` due to the new
dependency for `farchc`.

`libcommon-core` need to be both a `build-dependency` and a
`dependency` of `farchc` because the `_bindings_items.json` of
`libcommon-core` needs to be generated before running the build of
`farchc`.

Change-Id: I96853c37d3147d2309489abf6007e964b042af85
Priv-Id: 9c60b08c62e9e5921b0450a9b4d5b2f6a8665fe9