rust: allow to use a manual cargo package name.
It can happen that the rust package name can collide with the waf
targets already defined.
In that case, it is useful to be able to manually define the cargo
package name in the waf task gen that can be different from the waf
target.
By default, we still use the waf target name.

Change-Id: I3f19ea68426cd7e0fb9c383e99d38bb08c5b2acc
Refs: #112812
Priv-Id: 6904ac318f5169928df09328d03088509c6d656d