waf: Run ruff (from waf ruff and waf ruff-fix) from current directory.
Previously, ruff checks were always launched on the whole repository,
wherever the command is launched from. This might be a good idea for the
future, when mmsx will be 100% ruff-compliant, but right now it is
better to run the checks on the current directory only so that we can
migrate the codebase progressively.

Change-Id: Ie4842edeea698b24faf7462e0c525b11e6d98b2b
Priv-Id: 5627378fd5824694e3aba9c31fc40122063bfeb8