build: check validity of behave files in Z_LIST_RETRY env variable.
When zparser script detects an out of any test from a behave test, it
actually provides an invalid feature filename in the Z_LIST_RETRY
environment variable.

Even if the issue has been fixed on this side, this patch now enforces
the fact that a behave file/path with ".feature" is always expected, so
we can trigger the retry-behave/*.feature) in run_checks.sh bash
script... and not fall in another unexpected case!

Change-Id: Icc8fce7f3adc5c386d0827c67c4d25bab2cf0cb6
Priv-Id: 6ba850efd6bf785c8f9b9347c086edb41a7f2728