rust/waf: cargo target should be linked by cargo and not by waf.
Cargo targets should be linked by Cargo to other Cargo targets and not
by Waf.

A Rust dependency should always be a rust lib and linked by Cargo.

A Cargo package can be a "lib" and a "staticlib" at the same time
enabling using the same library for both Cargo targets and Waf targets.

Change-Id: Ia4ca04ed55fa104b0d7a4f81da6e96401469b60f
Refs: #112812
Priv-Id: 6b4daa64ee5053082a091cf08dd54da9282d6dd9