rust: support double pic for rust.
By default, we compile the targets in two modes, "pic" and "no-pic" for
performance reasons.

"no-pic" cannot be used to build shared libraries.

For Rust targets, we need to the same as C targets, duplicating the
compilation targets in Rust.

The way we can do it, is using another cargo profile to compile the
packages with "-pic".

Unfortunately, this means that we compile the external dependencies and
targets twice.

Change-Id: I8bcfe05fc0c8ad1f7a6e89d7d01e830e72cb1c45
Refs: #112812
Priv-Id: 3d5c412bdd3fc217767272e0eab1c3931c73f293