rust: move generated files to a `.waf-build` directory.
Instead of polluting the source directory with `_x` files, put them in a
single `.waf-build` directory in the source directory.

To avoid rewriting the same files in current and "pic" profile, separate
the "pic" profile generated files in a ".waf-build-pic" directory.

The binding file needs to be included with the new export environment
variable "BINDGEN_BINDING_FILE".

Change-Id: Id886a88df4ce088b2c4ea6cb8730ef7940c060f9
Refs: #112812
Priv-Id: a1fc779fb845287b9f4643a7f4bad496a7389c44