net/http: check connection state prior calling set_ready
Context:
Function `set_ready` is used to mark a connection as ready (in addition
with function `set_busy` that is used to mark connection as busy).

Function `set_ready` is first called when connection is established (or
when TLS handshake is done). During this first call, flag `is_connected`
is set to true.
Function `set_ready` is then called each time a HTTP query is done if it
was flagged as `busy` (and if it can be used to send another HTTP
query).

Problem:
When using a proxy and if server is using TLS, we first send a HTTP
CONNECT query to the proxy and we should not mark the connection as
ready until the TLS handshake is performed.

Solution:
Check flag `is_connected` that is set to true when connection is
established (or when TLS hanshake is done).

Change-Id: Ia09d76956ca2a85c4ec0356ff65107235872d7f3
Priv-Id: 6d9999fb71cf204ceee6bbcd28e0ded31bf9c869