net/http: handle CONNECT empty body response
Upon a successful CONNECT, HTTP server reply with HTTP 200 Connection
established without any body.
Without Content-Length header indication, HTTP parser was waiting for
a body as it didn't considered the query as completed (then the
connection was closed for inactivity).

Add a check on response `code` and `reason` to detect this case and set
the `chunk_length` to 0 so the response is considered as completed and
it will call `on_query_done` callback (similar for what is done with
code 204 - No Content).

Change-Id: Ia49592d0f4a2c1c3efef075448b2d1a59e092860
Priv-Id: cb0867da2b97277de8777b2f770941cd1c39796b