build: fix suite name issue on test failure.
This patch fixes and standardizes the name of the test suite for the
zparser, so it is always the same when entering and exiting execution of
tests. In fact, the FTK feature introduced on buildbot during the
hackathon took into consideration only the start of test but the test
failure case was forgotten.

Change-Id: If7ef0fe2eb3cebe5ad55725567dfa764bf9562f1
Priv-Id: c2b20745d03adac056f0f4f518209af618746021