iop-rpc-channel: handling of asynchronous address resolution.
Prior to this patch, address resolution of a remote peer inside
__ic_connect() could be done only in a synchronous manner despite the
O_NONBLOCK flag. As a consequence, this behavior could lead to the main
event loop not responding when answers from DNS providers were not
issued fast enough.

This patch now allows __ic_connect() calls to perform asynchronous
resolution of a remote address when the flag O_NONBLOCK is set *AND* the
thr module loaded before triggering the connection with the remote peer.
This way, the event loop of the process will be free to execute other
operations while waiting DNS response.

In case of ic wiped before the DNS response, a cancel flag is
positioned in the DNS context: callback later executed in the main
thread will just do nothing and wipe the result.

Change-Id: I1eed47384bae7fc181865fb2e86e785999feb316
Priv-Id: e535ea2bf449b1a68133280c1125d125af34cf9c