rust: explicitly list members of workspace instead of using '*'.
Cargo will try to consider every directories as a package even if the
directory is empty.
In that case, Cargo will produce an error saying it cannot find the
`Cargo.toml` of the package.
This cause issues when switching commits, where some empty directories
can remain in the repository.

The solution is to explicitly list the members instead of using '*'.
This is more painful but safer.

Change-Id: I676f4d9bffe94e7c8132b3d2d2d5a8b8badfd759
Refs: #112812
Priv-Id: 80fbea7aaa8c68ec815701a729318f4e14c5003f