rust: make sure IOP traits are dyn compatible.
To create and manipulate the IOPs in rust, the traits will need to by
dyn compatible.
This means they cannot be 'Sized' and they cannot contain const
variables.
See
https://doc.rust-lang.org/reference/items/traits.html#dyn-compatibility

To solve these issues, move the new() method to the derive
implementation and use a method to get the C description of structs and
unions.

Change-Id: I4a25e6f8a852a15207d6498cd90a172dbbd4ea0a
Refs: #112812
Priv-Id: 2e3ac021b6b1f620e7daa74d24e192dff3fbea61