rust: fix use libs for rust task gen.
To propagate the libs throught `uselib`, waf needs to know which
variables to propagate.
This is done by using `USELIB_VARS`.
By default, only the features known by waf are set in `USELIB_VARS`
(`cprogram`, `cshlib`, ...).

In order to make the rust targets to work the same as C targets, we need
to add the variables in `USELIB_VARS` for `rust`.

Change-Id: I8da848b942a2aa2c7cbc9cc88d5a52787b6a5e74
Refs: #112812
Priv-Id: 93717a705ccc904c9c3210e4f2f0dce2205a44c4