rust: fix sb buffer handler for SbStack.
Unfortunately, due to how Rust is moving and copying the value around,
having a pointer on an buffer on stack is complicated to do it
internally in the object.

Pin must be used and it is not easy to put it inside the structure that
is used by the user.
Instead, we should externalize the buffer, and pass it to SbStack.

A macro could be used to simplify the usage later on.

Change-Id: Ib79e63f6675f5b2dc1f902ef85cac42cb65c4106
Refs: #112812
Priv-Id: 3807d33392348eb9f6110c5d7ce6ba78476190f9