net-addr: add a way to resolve addresses asynchronously.
The current implementation of addr_info and addr_resolveXXX functions
are based on getaddrinfo() call to provide the resolution of an address.
As this call is currently done in a synchronous way, the event loop can
be stuck if this resolution takes too much time.

The purpose of this patch is to provide an asynchronous way to provide
the address resolution, with the help of:
 * thread job API already provided in lib-common (for more information,
   please read the "Base library" / "Thread jobs" section of the
   lib-common documentation) for the address to resolve.
 * a wake event object (event loop handler) to notify the end of the
   resolution to the main thread.

Two things have been taken into consideration for the implementation:
 * the module named "thr" must to be loaded before using the
   asynchronous calls. Otherwise, the job posted will never be launched.
 * the wake event object has been used in order to fix initial
   implementation using only thread job API. In fact, some pieces of
   code like the following one cannot work without a proper file
   descriptor during the asynchronous resolution.

   while (!_G.authenticated) {
       el_fd_loop(ic->elh, 1000, EV_FDLOOP_HANDLE_TIMERS);
   }

Associated tests related to these changes will be realized in the next
patch, so we can keep the git history of zchk-net.c file.

Change-Id: Ief1bab9a0fe62ea7a3864566d6c3c9c5c368a10f
Priv-Id: 05e0cb3f21d4d214e14a0db7631f3c49fb7f8e08