iop/env: fix unregistering struct from RPC on unregister pkg.
Since d5cfad4c30c, we register and unregister the structs created for
the RPCs.
The test that the registered struct was done specially for the RPC was
buggy and did not actually matched the structs (a '.' skip was missing).
This caused the structs to not actually be unregistered, and can
conflict when another DSO with the same "symbol" was registered again.

Fix it by skipping the dot, and add a test.

Change-Id: I9e74dcfefdfe22fae14a7f26c89d46cb99f6668b
Priv-Id: 6d29b5cf3c46940997a52ab9de4112866bc106e3