iop: stop casting arrays to lstr_t and fix uninitialized memory
There was a lot of places where the IOP arrays were cast cast to lstr_t
to be manipulated in a generic way regardless of the actual type of
array. If it historically works because the lstr_t type and all the IOP
arrays share the same structure layout it's still an awful practice and
it really doesn't help the code readability.

This patch adds the type iop_array_void_t and uses it everywhere an
lstr_t was used to access an IOP array. It also fixes a few places were
the `flags` member of the IOP array were left uninitialized. It's not a
big deal since this member is actually unused and exists because of the
layout compatibility with the lstr_t, but Valgrind is not happy about
it.

Change-Id: Ib6dc79954fd8359ac60b2072109722e853137c8d
Priv-Id: a533d16857334c98ac1ace9629d0be283ea85621