cflags: remove '-Wdeclaration-after-statement'.
In our coding rules we specify that variable declarations need to be
done before statements.

Before, this was enforced by '-Wdeclaration-after-statement' warning.

Unfortunately, this blocks us to use some interesting constructs and
macros like "defer" which we will introduce in the next commit.

So let's remove this warning in the cflags.

It is still enforced in our coding rules.

Change-Id: Ie8b7b718843737efd514e6f0ae6568ff16a7a105
Priv-Id: 524943ea5778dc2b185ce87e9d1a7a9e961b71fa