core-macros: add 'defer' implementation.
Inspired by Go `defer` statements, this macro can be used to defer the
execution at the end of the scope.

Internally, it uses:
- For GCC: __attribute__((cleanup)) and nested functions.
- For Clang: __attribute__((cleanup)) and blocks (with `-fblocks`).

In order to support both compilers, the variables used on the defer
blocks *MUST* be declared with `deferred` keyword and considered passed
by reference (C++ reference style) in the block.
Failure to do so can result in some undesirable consequences.

The defer block acts as a separate function. Using `return` in the defer
block only returns in the block itself and not in the original function.

Example:
----
 void foo(void)
 {
     deferred qv_t(i32) qv

     qv_init(&qv);
     defer({
         qv_wipe(&qv);
     });

     ... Use the qv without qv_wipe() at the end.
 }
----

Change-Id: I222396cf3d8ac4189b0a1636f847daa4ff661b0d
Priv-Id: fe2dd028bcac6667d015a9415ede25f1bc56398d