thr-jobs: handle cgroups limits
With cgroups, we can have a limit on the number of CPU or CPU time
allowed to the process. This limit would have an impact on the thread
jobs library, but it was not handled. The module relied only on the
number of CPUs seen by the OS which does not handle cgroups limits.

This patch retrieve and apply those limits. We can have 2 kinds of
limits:
- limit on CPU usable.
- limit on CPU time (period and quota).

Both limits are handled. The thread job module take the maximum of the
number of CPU and cgroups CPU number limit. Then period and quota are
applied. If the quota/period ratio is greater than 1, it is applied as
a number of CPU. Otherwise, the ratio is applied to the previously
computed number of CPU.

Change-Id: I5e525b964e2c100d0a79524e498da6d873e02299
Priv-Id: 3eef2926cf3d655cdbcf7d93e97428c022326df1