static-checks: add "staged_files" to only check staged files
It will be useful to have a git hook use static-checks.sh.
My code contains a comment dedicated to shellcheck, I know it's not yet
commonly used in Intersec but it's the best tool to check shell/bash so
I use it and I hope other people will use it too.
--force-exclude to make it respect the excluded files of ruff.

Change-Id: I8c215a0fe027a5576fe1aadd78ade495747d6945
Priv-Id: ce158e5a1ff6c542b0f6bf290d242cb04d283e18