z: Make description of Z_TEST optional.
Most of time, the description was left empty or filled with the same
string as the name, or a useless rephrasing of it.

Make it optional, so that we can omit it in most cases.

Also remove the description of all tests were it has no additional value
compared to the name.

Change-Id: Idae07885389c83ad9e2d03caa78ce4c1c80da823
Priv-Id: f141b6b2f1131156d2bf3bd897213bf2d05f7948