thr-job: allow to run thr_syn_wait without draining the main queue
In some cases we want to launch several jobs and collect them before
executing anything else on the main thread. Currently when waiting for
the jobs to finish with thr_syn_wait() the main thread will schedule any
job awaiting to be handled on the main queue while waiting for the
workers to finish.

This patch introduces the function thr_syn_wait_flags() and the flag
THR_SYN_WAIT_NO_MAIN_DRAIN to forbid thr_syn_wait() to consume jobs from
the main queue.

Change-Id: I471d9b1248bf6897bd3eb052c49c75db55064906
Priv-Id: 1e10ca70c5a9de0fc18e05eb4d2fe4f85124cc22