bit-wah: use a true constant in release instead of _G.bits_in_bucket
The number of bits in a wah bucket was stored in a global variable to
allow the tests to play with it. But this came with a non negligible
cost in performance.

Since in release build we don't need at all to change this value, we now
use a true constant in P=release.

With this patch we observed up to 30% of additional performances during
big slot merges in QTS.

Change-Id: I00195c91948c6a51ad3ae8364ec6340ccb4d9da0
Priv-Id: c6eb861e0fda7ee76c2924800170ffa83c736092