bit-wah: remove the size hint on t_wah_new()
t_wah_new() used to take a size hint on the first internal bucket of the
WAH. This parameter requires a deep understanding on the WAH internals
and should not be exposed.

The only legitimate (but debatable) usage in wah_add1_at() will be
made obsolete by a rework of the buckets in a following patch anyway.

Change-Id: Ice1316161d7a077ad8a7d5fa2cb9af6c7693c744
Priv-Id: ec91ee2862ebb1a6e775ba7d0fb453379814f84c