bit-wah: improve wah_copy() implementation
When copying a WAH from source to dest we either shrunk or grown the
buckets of the dest WAH to match the size of the source WAH. Both
operation were done one bucket by one bucket, which resulted in a lot of
small reallocations of the buckets for big WAHs.

In addition, because the code was using the generic qvector APIs, every
destination buckets ended with an additional amount of pre-allocated
memory which is wasted for already closed buckets.

This patch brings several improvements:
    - memory reallocation are now done in a single call;
    - the last bucket of the destination WAH is preserved as the last
      bucket since it's supposed to be the only one with extra memory;
    - when a closed destination bucket needs to be grown we allocate the
      exact amount of required memory;
    - memcpy() is used instead of memmove() since the source and
      destination WAH aren't supposed to overlap.

Around +20% of performance was observed during a big merge of QTS slots
with this patch.

Change-Id: I64efd262f7c0b3b9facd05ec43cad960ccc10298
Priv-Id: 5d2897a22eb54ce2fb6c6b9ad988f738eeac5d68