bit-wah: improve memory footprint of small buckets
A WAH bucket used to be directly implemented as a qv_t of wah words. So
in sparse WAH we have dozen of very small allocated buckets (usually 2
words to indicate a range full of 0 or full of 1). In big WAH we're
talking about a hundred of buckets resulting in indirect memory access
when scanning the WAH for nothing.

Now sizeof(qv_t) is 24 bytes, which makes 5 WAH words plus a length
which is far sufficient to store long run of the same bit.

So this patch introduces a more complex type (union) to represent a
bucket (wah_bucket_t) which can be either an “inlined” bucket,
containing up to 5 words or a qv_t.

All buckets start as an “inlined“ bucket and then are upgrade to a qv_t
when needed.

This patch improved the speed of some big merges in QTS approximately by
a factor x3.

Change-Id: Iaa46092156c7e77bfefb2c0891501c236b0934c9
Priv-Id: 0a068848a76d31dfdae82ead06e24f779de38e4d