rust/farch: fix and rework buffer handling for farch.
Instead of using Vec<u8> to handle the buffers, use the t_scope to avoid
copying the buffers multiple times.
Use MaybeUninit to avoid unnecessary initializations, but it makes the
buffer handling more complicated (the helpers to handle slices of
MaybeUninit are unstable in Rust).

Fix and test the null character at the end of the buffers as it is
required for some part of our products (we sometime use `lstr.s`
directly :( ).

Change-Id: I0cf765df0f0148230dc22c57bd2b41dec8496a4e
Priv-Id: 1c9bda30eb6c45dc65ff4484c4ab0674380fd058