bit-wah: pack all runs of the same bit in the first bucket
When dealing with high bit ranges, the WAH often start with a lot of
small buckets sharing the same bit value. Those buckets are inefficient
both at packing bits and in term of performances.

This patch ensures that as long as one adds the same bit value in a WAH
they will all be packet in the first bucket, allowing maximum
compression (wah_word_t.count gets maximized) and avoiding the creation
of useless buckets.

As soon as the first bit of different value than the previous ones is
added in the WAH the packing stops and the buckets partition is resumed
normally.

This patch improved the speed of some big merges in QTS approximately by
a factor x6.

Change-Id: Icf21aaf0053f25cf2cb23a0a705901fcf15ddb98
Priv-Id: 7c9d0e33969013f6222eb5ad6f2e0c40d43aebd3