iop-ts: add `declare` for public type.
With newer version of TypeScript, when using class, we cannot redefine
the type of an attribute in an inherited class.
If the redefinition is expected, we have to declare it with `declare`.

For example,
```
class Alpha {
  public value: number;
  public struct: {
    type: 'number';
  };
}

class Beta extends Alpha {
  public text: string;
  public struct: Alpha['struct'] & {
    length: number;
  };
}
```
The following error is prompted "
Property 'struct' will overwrite the base property in 'Alpha'. If this
is intentional, add an initializer. Otherwise, add a 'declare' modifier
or remove the redundant declaration.ts(2612)
"

To avoid such issue and because we want to redefine the attribute, we
must use `declare`.

This commit add a `declare` before every public attributes.

Change-Id: I31fd2ec5dc397d7697415725024a74083b001c85
Priv-Id: 19bef1429e4e7073f8a6e5716c72dee39003c8ea