bit-wah: fix loading of WAH exactly aligned on the size of a bucket
A valid WAH must have a last_run_pos greater or equal to zero since it
always have at least a first bucket with two empty words (in which case
last_run_pos equals zero).

Yet the parser loading a serialized WAH starts with a last_run_pos at -1
since the first bucket is not created when starting to parse. This is
not an issue since any valid WAH will be serialized with at least two
words and thus the first bucket will be immediately created and
last_run_pos will be fixed.

However, after the completion of each bucket, the parser is reset to its
“starting” state, setting last_run_pos to -1 again. This is broken if
the parsing ends  immediately, because instead of having a last_run_pos
pointing on the last run of the last bucket we end with -1 which aborts
in wah_check_invariant() in non release build.

So the fix is to not reset last_run_pos/previous_run_pos in case we are
at then end of the serialized payload.

Change-Id: Icedb6fb61f0d8e9a1aa3f93849dbcb037501f30d
Priv-Id: 54b381b20fe25e061bec0aa0b13443580c70b29e