bit-wah: fix enumeration of sub-optimal WAH
In some rare situation the encoding of a WAH can be sub-optimal. For
example it can be the case of WAH loaded with too many bits per buckets
and split with from_data_split_chunk().

Let's consider the following optimal WAH composed of 1 run of 3 words:

    { words: 3, bit: 0 }, { count: 1 }, { literal: 0x12345678 }

This WAH contained 96 bits set to 0 and the word 0x12345678 can be
equally rewritten as:

    { words: 3, bit: 0}, { count: 0},
    { words: 0, bit: 0}, { count: 1}, { literal: 0x12345678 }

This WAH contains the same information but is composed of 2 runs for a
total of 5 words.

Now the wah_word_enum_next() function didn't covered the transition from
a run without literal to a run exclusively composed of literals. This
transition caused and random move in the current bucket, causing random
behaviors.

Change-Id: I2e5ef614b07980a0a35c33d4fa445710c2121221
Priv-Id: e670511d1e3e230e9a4db70a35100f76cced596a