rust: add 'deps-workspace-hack' to speed up the compilation.
In order to avoid compiling the same dependencies crate with some
feature differences because of dependency hierarchy, use a crate
`deps-workspace-hack` that list all the workspace dependencies and is
always used when running `cargo build` by waf.

This does not mean that in the result compilation binaries there will be
all the dependencies of the workspace. Only the dependencies of the
crate are still used for a given crate. But the dependencies will use
all the features specified in the workspaces.

Change-Id: I0a5ce6805696ad4631cd947e05f7067ca6af4a98
Priv-Id: 5c89f43878f59e48355ffea5eb24cc1c53e3ca08