str-buf-quoting: Fix escaped $ in sb_add_expandenv.
The documentation of sb_add_expandenv says that to have a '$' in the
output string, it must be escaped with '\'. Which was true, but the
problem is that everything that followed was removed!

This is because of a break used instead of a continue in the code.

The must funny thing is that there was a test of '\$'... without
anything after. Just like if the developper writing the test saw the
bug, and was too lazy to fix it.

Change-Id: I8273325e43027fe69a0a8f7b5a266a3db8d89ce9
Priv-Id: 3dd900a6cf61a9293be47ca08119c34d5b62327d