bit-wah: fix loading of wah generated before the overfill feature
A WAH that was generated without bucket overfilling cannot be loaded
with applying an overfilling of the first bucket.

Indeed, overfilling the first bucket changes the distribution of the
chunks inside the following buckets which may not stay aligned on
WAH_BITS_IN_BUCKET. And since the loading is read-only the content of
the chunks cannot be modified.

So the solution is to detect if the WAH was already encoded with the
bucket overfilling or not before loading it by looking at the first
bucket.

Change-Id: I55d3005a9053319af0c2fc8b710ea254aa277b2a
Priv-Id: b6b0c6acdc50cd0382b369159f713fbcc7640c28