build/rust: fix target build dir for debug + pic.
When using PIC, the compilation is done with two cargo profiles, non-pic
and pic.
This results on two target profile directories.
But there is something special for debug.
'debug' waf profile uses the 'dev' cargo profile, but it lands in the
'debug' target cargo directory.
But for 'pic', it lands in the 'dev-pic' cargo directory.

In the other waf profiles, the cargo target directories have the same
name as the waf profile, e.g. 'asan' and 'asan-pic'.

We need to add a special handling for the debug profile.

Change-Id: I588b2943582ea8979474e030a0e2e3b5e6c6f16d
Priv-Id: aa375bb7f986746ea9c8c1407f0f2eb59ba6ca97