core/mem: fix assert on zero size alloc when using "BYPASS_MEMPOOL".
When doing a zero size allocation, our allocators do not return NULL,
but MEM_EMPTY_ALLOC since 99ebbe55918c.

An assert was added to check that we well return MEM_EMPTY_ALLOC when
doing a zero size allocation.

At least, this is what it is expected in normal cases, but there is one
special case where we don't do that, it is when using `BYPASS_MEMPOOL=1`.
In that case, an header is always added to the allocation when using
stack frames, and the final allocation size is never zero even when
requesting an initial zero size allocation.

This triggers the assert checking for `MEM_EMPTY_ALLOC` because we
actually have something allocated.

However, the whole point of the assert was to check that the allocator
did not return NULL, and not particularly `MEM_EMPTY_ALLOC`.

So we can modify the assert to check that we returned something and not
NULL, and it works in all cases.

Change-Id: I5a9a6680e68868d7f4a665ffe3da4a836bfde0ee
Priv-Id: 3c013568374d55aa95078c837a32417dc46f3620