iop/http: allow to retrieve the http channel of a query
The `http_iop_channel_from_msg` helper allows to retrieve the
`http_iop_channel_t` associated to an `http_iop_msg_t`. This is
especially useful in the reply callback.

However to be able to do this, the generic HTTP client needed to stop
resetting the query owner before calling the query callback. The
historical code was weird on this point because it led to think that the
`owner` of a query could be NULL, but AFAIK it's not possible, at least
not in the code paths leading to `httpc_query_on_done`.

To avoid leaving a dandling pointer in the query `owner` we reset it in
`httpc_query_reset` which is sufficient since at the end of the
`on_done` callback the user is supposed to either reset or destroy the
query, it cannot be reused as is.

Change-Id: I129e4226541bbc88d4c6ddb45212319cffd7a7ae
Priv-Id: 4021c534f654be29877a63d604c79614e1c76b80