iop/http: support pending queries in http_iop_channel_from_msg
Before being attached to an `httpc_t` connection, the `http_iop_msg_t`
could be hold in the pending queue of the `http_iop_channel_t` in which
case the `http_iop_channel_from_msg` was wrong.

So we add a `priv_owner` member in a union aliasing the `owner` member
of the `httpc_query_t`, allowing users to store a “private” owner until
the query gets actually attached to an `httpc_t`.

Change-Id: I4c4488ac4e6c1c5f68833a7eb6e6846d101da23e
Priv-Id: 2dc27b5402fcb23bae4d402d0861890f0b03a5bb