httpc: fix pool detach done too early
The HTTPC pool detach was performed before canceling the pending queries
and thus we lost the link between the HTTP connection and its pool in
those callbacks.

We cannot perform the pool detach after canceling the pending queries
either because we could recurse if the query callback decided to close
the pool.

So we have to detach the pool in two step, we first remove the
connection from the pool itself (no more risk of recursion), then we
cancel the queries and finally we remove the pool information from the
connection.

Change-Id: I0bb77ffc2298783f3b7ef073581f5c50f795e0b3
Priv-Id: de2d9c984091bdcd9dee6164a2a0b31a1be08b08